+++ workflow/scripts/csaw/csaw_readparam.R
@@ -13,10 +13,21 @@ sink(log_file, type = "message")
 
 # Load libraries
 base::library(package = "csaw", character.only = TRUE)
+base::library(package = "rtracklayer", character.only = TRUE)
 base::message("Libraries loaded")
 
+standard_chrom <- c(1:23, "X", "Y")
+if ("organism" %in% base::names(x = snakemake@params)) {
+    if (base::as.character(x = snakemake@params[["organism"]]) == "mm10") {
+        standard_chrom <- c(1:19, "X", "Y")
+    }
+}
+
+blacklist <- base::as.character(x = snakemake@input[["blacklist"]])
+blacklist <- rtracklayer::import(blacklist)
+
 # Build command line
-extra <- "minq=20"
+extra <- "minq=20, restrict=standard_chrom, discard=blacklist"
 if ("extra" %in% base::names(x = snakemake@params)) {
     extra <- base::as.character(x = snakemake@params[["extra"]])
 }
@@ -38,6 +49,7 @@ base::saveRDS(
     object = read_params,
     file = snakemake@output[["rds"]]
 )
+base::message("RDS saved, process over")
 
 # Proper syntax to close the connection for the log file
 # but could be optional for Snakemake wrapper